Capture TextEditor lint messages. Fix lint message clearing on file move/copy/etc.
